+++ src/azure-cli/azure/cli/command_modules/network/custom.py
@@ -2058,11 +2058,163 @@ def list_express_route_ports(cmd, resource_group_name=None):
 
 
 # region PrivateEndpoint
+def create_private_endpoint(cmd, resource_group_name, private_endpoint_name, subnet,
+                            private_connection_resource_id, connection_name, group_ids=None,
+                            virtual_network_name=None, tags=None, location=None,
+                            request_message=None, manual_request=None):
+    client = network_client_factory(cmd.cli_ctx).private_endpoints
+    PrivateEndpoint, Subnet, PrivateLinkServiceConnection = cmd.get_models('PrivateEndpoint',
+                                                                           'Subnet',
+                                                                           'PrivateLinkServiceConnection')
+    pls_connection = PrivateLinkServiceConnection(private_link_service_id=private_connection_resource_id,
+                                                  group_ids=group_ids,
+                                                  request_message=request_message,
+                                                  name=connection_name)
+    private_endpoint = PrivateEndpoint(
+        location=location,
+        tags=tags,
+        subnet=Subnet(id=subnet)
+    )
+
+    if manual_request:
+        private_endpoint.manual_private_link_service_connections = [pls_connection]
+    else:
+        private_endpoint.private_link_service_connections = [pls_connection]
+
+    return client.create_or_update(resource_group_name, private_endpoint_name, private_endpoint)
+
+
+def update_private_endpoint(instance, cmd, tags=None, request_message=None):
+    with cmd.update_context(instance) as c:
+        c.set_param('tags', tags)
+
+    if request_message is not None:
+        if instance.private_link_service_connections:
+            instance.private_link_service_connections[0].request_message = request_message
+        else:
+            instance.manual_private_link_service_connections[0].request_message = request_message
+
+    return instance
+
+
 def list_private_endpoints(cmd, resource_group_name=None):
     client = network_client_factory(cmd.cli_ctx).private_endpoints
     if resource_group_name:
         return client.list(resource_group_name)
     return client.list_by_subscription()
+# endregion
+
+
+# region PrivateLinkService
+def create_private_link_service(cmd, resource_group_name, service_name, subnet, frontend_ip_configurations,
+                                private_ip_address=None, private_ip_allocation_method=None,
+                                private_ip_address_version=None,
+                                virtual_network_name=None, public_ip_address=None,
+                                location=None, tags=None, load_balancer_name=None,
+                                visibility=None, auto_approval=None, fqdns=None):
+    client = network_client_factory(cmd.cli_ctx).private_link_services
+    FrontendIPConfiguration, PrivateLinkService, PrivateLinkServiceIpConfiguration, PublicIPAddress, Subnet = \
+        cmd.get_models('FrontendIPConfiguration', 'PrivateLinkService', 'PrivateLinkServiceIpConfiguration',
+                       'PublicIPAddress', 'Subnet')
+    pls_ip_config = PrivateLinkServiceIpConfiguration(
+        name='{}_ipconfig_0'.format(service_name),
+        private_ip_address=private_ip_address,
+        private_ip_allocation_method=private_ip_allocation_method,
+        private_ip_address_version=private_ip_address_version,
+        subnet=subnet and Subnet(id=subnet),
+        public_ip_address=public_ip_address and PublicIPAddress(id=public_ip_address)
+    )
+    link_service = PrivateLinkService(
+        location=location,
+        load_balancer_frontend_ip_configurations=frontend_ip_configurations and [
+            FrontendIPConfiguration(id=ip_config) for ip_config in frontend_ip_configurations
+        ],
+        ip_configurations=[pls_ip_config],
+        visbility=visibility,
+        auto_approval=auto_approval,
+        fqdns=fqdns,
+        tags=tags
+    )
+    return client.create_or_update(resource_group_name, service_name, link_service)
+
+
+def update_private_link_service(instance, cmd, tags=None, frontend_ip_configurations=None, load_balancer_name=None,
+                                visibility=None, auto_approval=None, fqdns=None):
+    FrontendIPConfiguration = cmd.get_models('FrontendIPConfiguration')
+    with cmd.update_context(instance) as c:
+        c.set_param('tags', tags)
+        c.set_param('load_balancer_frontend_ip_configurations', frontend_ip_configurations and [
+            FrontendIPConfiguration(id=ip_config) for ip_config in frontend_ip_configurations
+        ])
+        c.set_param('visibility', visibility)
+        c.set_param('auto_approval', auto_approval)
+        c.set_param('fqdns', fqdns)
+    return instance
+
+
+def list_private_link_services(cmd, resource_group_name=None):
+    client = network_client_factory(cmd.cli_ctx).private_link_services
+    if resource_group_name:
+        return client.list(resource_group_name)
+    return client.list_by_subscription()
+
+
+def update_private_endpoint_connection(cmd, resource_group_name, service_name, pe_connection_name,
+                                       connection_status, description=None, action_required=None):
+    client = network_client_factory(cmd.cli_ctx).private_link_services
+    PrivateEndpointConnection, PrivateLinkServiceConnectionState = cmd.get_models('PrivateEndpointConnection',
+                                                                                  'PrivateLinkServiceConnectionState')
+    connection_state = PrivateLinkServiceConnectionState(
+        status=connection_status,
+        description=description,
+        action_required=action_required
+    )
+    pe_connection = PrivateEndpointConnection(
+        private_link_service_connection_state=connection_state
+    )
+    return client.update_private_endpoint_connection(resource_group_name, service_name, pe_connection_name, pe_connection)  # pylint: disable=line-too-long
+
+
+def add_private_link_services_ipconfig(cmd, resource_group_name, service_name,
+                                       private_ip_address=None, private_ip_allocation_method=None,
+                                       private_ip_address_version=None,
+                                       subnet=None, virtual_network_name=None, public_ip_address=None):
+    client = network_client_factory(cmd.cli_ctx).private_link_services
+    PrivateLinkServiceIpConfiguration, PublicIPAddress, Subnet = cmd.get_models('PrivateLinkServiceIpConfiguration',
+                                                                                'PublicIPAddress',
+                                                                                'Subnet')
+    link_service = client.get(resource_group_name, service_name)
+    if link_service is None:
+        raise CLIError("Private link service should be existed. Please create it first.")
+    ip_name_index = len(link_service.ip_configurations)
+    ip_config = PrivateLinkServiceIpConfiguration(
+        name='{0}_ipconfig_{1}'.format(service_name, ip_name_index),
+        private_ip_address=private_ip_address,
+        private_ip_allocation_method=private_ip_allocation_method,
+        private_ip_address_version=private_ip_address_version,
+        subnet=subnet and Subnet(id=subnet),
+        public_ip_address=public_ip_address and PublicIPAddress(id=public_ip_address)
+    )
+    link_service.ip_configurations.append(ip_config)
+    return client.create_or_update(resource_group_name, service_name, link_service)
+
+
+def remove_private_link_services_ipconfig(cmd, resource_group_name, service_name, ip_config_name):
+    client = network_client_factory(cmd.cli_ctx).private_link_services
+    link_service = client.get(resource_group_name, service_name)
+    if link_service is None:
+        raise CLIError("Private link service should be existed. Please create it first.")
+    ip_config = None
+    for item in link_service.ip_configurations:
+        if item.name == ip_config_name:
+            ip_config = item
+            break
+    if ip_config is None:  # pylint: disable=no-else-return
+        logger.warning("%s ip configuration doesn't exist", ip_config_name)
+        return link_service
+    else:
+        link_service.ip_configurations.remove(ip_config)
+        return client.create_or_update(resource_group_name, service_name, link_service)
 # endregion
 
 
@@ -2074,7 +2226,8 @@ def create_load_balancer(cmd, load_balancer_name, resource_group_name, location=
                          public_ip_dns_name=None, subnet=None, subnet_address_prefix='10.0.0.0/24',
                          virtual_network_name=None, vnet_address_prefix='10.0.0.0/16',
                          public_ip_address_type=None, subnet_type=None, validate=False,
-                         no_wait=False, sku=None, frontend_ip_zone=None, public_ip_zone=None):
+                         no_wait=False, sku=None, frontend_ip_zone=None, public_ip_zone=None,
+                         private_ip_address_version=None):
     from azure.cli.core.util import random_string
     from azure.cli.core.commands.arm import ArmTemplateBuilder
     from azure.cli.command_modules.network._template_builder import (
@@ -2125,7 +2278,7 @@ def create_load_balancer(cmd, load_balancer_name, resource_group_name, location=
     load_balancer_resource = build_load_balancer_resource(
         cmd, load_balancer_name, location, tags, backend_pool_name, frontend_ip_name,
         public_ip_id, subnet_id, private_ip_address, private_ip_allocation, sku,
-        frontend_ip_zone)
+        frontend_ip_zone, private_ip_address_version)
     load_balancer_resource['dependsOn'] = lb_dependencies
     master_template.add_resource(load_balancer_resource)
     master_template.add_output('loadBalancer', load_balancer_name, output_type='object')
@@ -2237,7 +2390,7 @@ def set_lb_inbound_nat_pool(
 def create_lb_frontend_ip_configuration(
         cmd, resource_group_name, load_balancer_name, item_name, public_ip_address=None,
         public_ip_prefix=None, subnet=None, virtual_network_name=None, private_ip_address=None,
-        private_ip_address_allocation=None, zone=None):
+        private_ip_address_version=None, private_ip_address_allocation=None, zone=None):
     FrontendIPConfiguration, SubResource, Subnet = cmd.get_models(
         'FrontendIPConfiguration', 'SubResource', 'Subnet')
     ncf = network_client_factory(cmd.cli_ctx)
@@ -2249,6 +2402,7 @@ def create_lb_frontend_ip_configuration(
     new_config = FrontendIPConfiguration(
         name=item_name,
         private_ip_address=private_ip_address,
+        private_ip_address_version=private_ip_address_version,
         private_ip_allocation_method=private_ip_address_allocation,
         public_ip_address=SubResource(id=public_ip_address) if public_ip_address else None,
         public_ip_prefix=SubResource(id=public_ip_prefix) if public_ip_prefix else None,
@@ -2264,15 +2418,19 @@ def create_lb_frontend_ip_configuration(
 
 def set_lb_frontend_ip_configuration(
         cmd, instance, parent, item_name, private_ip_address=None,
-        private_ip_address_allocation=None, public_ip_address=None, subnet=None,
-        virtual_network_name=None, public_ip_prefix=None):
+        private_ip_address_allocation=None, public_ip_address=None,
+        subnet=None, virtual_network_name=None, public_ip_prefix=None):
     PublicIPAddress, Subnet, SubResource = cmd.get_models('PublicIPAddress', 'Subnet', 'SubResource')
-    if private_ip_address == '':
+    if not private_ip_address:
         instance.private_ip_allocation_method = 'dynamic'
         instance.private_ip_address = None
     elif private_ip_address is not None:
         instance.private_ip_allocation_method = 'static'
         instance.private_ip_address = private_ip_address
+
+    # Doesn't support update operation for now
+    # if cmd.supported_api_version(min_api='2019-04-01'):
+    #    instance.private_ip_address_version = private_ip_address_version
 
     if subnet == '':
         instance.subnet = None
@@ -3658,7 +3816,8 @@ def create_subnet(cmd, resource_group_name, virtual_network_name, subnet_name,
 
 def update_subnet(cmd, instance, resource_group_name, address_prefix=None, network_security_group=None,
                   route_table=None, service_endpoints=None, delegations=None, nat_gateway=None,
-                  service_endpoint_policy=None):
+                  service_endpoint_policy=None, disable_private_endpoint_network_policies=None,
+                  disable_private_link_service_network_policies=None):
     NetworkSecurityGroup, ServiceEndpoint, SubResource = cmd.get_models(
         'NetworkSecurityGroup', 'ServiceEndpointPropertiesFormat', 'SubResource')
 
@@ -3697,6 +3856,12 @@ def update_subnet(cmd, instance, resource_group_name, address_prefix=None, netwo
 
     if delegations:
         instance.delegations = delegations
+
+    if disable_private_endpoint_network_policies is not None:
+        instance.private_endpoint_network_policies = disable_private_endpoint_network_policies
+
+    if disable_private_link_service_network_policies is not None:
+        instance.private_link_service_network_policies = disable_private_link_service_network_policies
 
     return instance