+++ src/azure-cli/azure/cli/command_modules/role/custom.py
@@ -779,7 +779,7 @@ def create_application(cmd, display_name, homepage=None, identifier_uris=None,  
         result = graph_client.applications.create(app_create_param)
     except GraphErrorException as ex:
         if 'insufficient privileges' in str(ex).lower():
-            link = 'https://docs.microsoft.com/en-us/azure/azure-resource-manager/resource-group-create-service-principal-portal'  # pylint: disable=line-too-long
+            link = 'https://docs.microsoft.com/azure/azure-resource-manager/resource-group-create-service-principal-portal'  # pylint: disable=line-too-long
             raise CLIError("Directory permission is needed for the current user to register the application. "
                            "For how to configure, please refer '{}'. Original error: {}".format(link, ex))
         raise
@@ -974,6 +974,8 @@ def update_application(instance, display_name=None, homepage=None,  # pylint: di
         app_patch_param.identifier_uris = identifier_uris
     if reply_urls is not None:
         app_patch_param.reply_urls = reply_urls
+    if homepage is not None:
+        app_patch_param.homepage = homepage
 
     return app_patch_param